requirements.txt
anyio==3.6.1
autopep8==1.7.0
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
dnspython==2.2.1
fastapi==0.85.0
graphql-core==3.2.3
h11==0.14.0
httptools==0.5.0
idna==3.4
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.2
Pygments==2.13.0
PyJWT==1.7.1
pymongo==4.2.0
python-dateutil==2.8.2
python-decouple==3.3
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
strawberry-graphql==0.132.1
toml==0.10.2
typing_extensions==4.3.0
uvicorn==0.18.3
watchfiles==0.17.0
websockets==10.3
